requirements.txt
aiosqlite==0.17.0
alembic==1.8.1
annotated-types==0.7.0
anyio==4.9.0
asyncpg==0.27.0
bcrypt==4.0.1
beautifulsoup4==4.13.4
certifi==2025.6.15
cffi==1.15.1
charset-normalizer==3.4.2
click==8.2.1
cryptography==38.0.1
dbus-python==1.2.18
dnspython==2.2.1
ecdsa==0.18.0
email-validator==1.3.0
fastapi==0.115.14
finance-datareader==0.9.96
greenlet==1.1.3.post0
h11==0.16.0
httpcore==1.0.9
httptools==0.5.0
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
lxml==6.0.0
Mako==1.2.3
MarkupSafe==3.0.2
narwhals==1.45.0
numpy==2.3.1
packaging==25.0
pandas==2.3.0
passlib==1.7.4
plotly==6.2.0
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycairo==1.28.0
pycparser==2.21
pycryptodome==3.23.0
pydantic==1.10.2
PyGObject==3.42.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-jose==3.3.0
python-multipart==0.0.5
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-file==2.1.0
rsa==4.9
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
SQLAlchemy==2.0.12
starlette==0.46.2
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
uvloop==0.17.0
watchfiles==1.1.0
websockets==15.0.1
